pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "A Discord bot for Wowhead search"
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"

dependencies = [
    "discord.py",
    "requests",
    "python-dotenv",
    "beautifulsoup4"
]

